+++ server.js
@@ -17,7 +17,7 @@ const readerStreamsRoute = require('./routes/deprecated/reader-streams')
 const readerRoute = require('./routes/deprecated/reader')
 const whoamiRoute = require('./routes/whoami')
 const inboxRoute = require('./routes/deprecated/inbox')
-const readersRoute = require('./routes/readers')
+const readersRoute = require('./routes/reader-post')
 const getOutboxRoute = require('./routes/deprecated/outbox-get')
 const postOutboxRoute = require('./routes/deprecated/outbox-post')
 const fileUploadRoute = require('./routes/file-upload')
@@ -37,11 +37,11 @@ const publicationDeleteRoute = require('./routes/publication-delete')
 const publicationPutTagRoute = require('./routes/publication-put-tag')
 const publicationDeleteTagRoute = require('./routes/publication-delete-tag')
 const publicationGetRoute = require('./routes/publication-get')
-const readerLibraryRoute = require('./routes/reader-library')
+const readerLibraryRoute = require('./routes/library-get')
 const tagPostRoute = require('./routes/tag-post')
 const tagPatchRoute = require('./routes/tag-patch')
 const tagDeleteRoute = require('./routes/tag-delete')
-const readerNotesRoute = require('./routes/reader-notes')
+const readerNotesRoute = require('./routes/readerNotes-get')
 const getJobRoute = require('./routes/job-get')
 const getNoteRoute = require('./routes/note-get')
